feat: added newline command to snippets that require newlines
This commit adds the newline (`&`) when it makes sense for a snippet to be
written on a newline. For example org src-blocks and headers need to be on their
own lines, so if a snippet is called in the middle of a line this will ensure
that those snippets are correctly written on the next line.